Add files via upload
python script : compute the number of pages, based on the number of files in each directory (one directory = a Cahier) ; using the datas in the output of the function parsezip_and_count (for each .xml, there is a jpeg2, and there is one pdf in the directory => (total of file - 1) / 2). output a dataframe pour all the corpus (part1 and 2 concatenated).